Added support for external document root setting
These changes allow an external SLIRConfig::$documentRoot (e.g.
"http://example.com") to be used in the config. In the slirrequest class
also external files are checked before an "Image does not exist" message
is sent back. In slir class the last modified date is fetched via http
if the full path is a valid url.